Make vim plugin use importjs binary
Now that we have rewritten the importjs binary in Node, we are starting
to port over editor plugins one by one. The vim plugin used to require
Ruby to function. That requirement is now gone, and it's written in
"pure" vimscript.

Because importjs outputs and expects json, we need a json parser and
encoder in vim. Prior to 7.4.1304, the json_encode and json_decode
functions didn't exist. After talking to @wincent about it, I decided to
keep that as a requirement (having a vim equal to or newer than
7.4.1304). If that turns out to be a problem, we can allow people to use
the WebAPI plugin [1], which has both json decoding and encoding built
in. Before using built-in functions (json_encode and json_decode), I
successfully used WebAPI.

[1]: http://www.vim.org/scripts/script.php?script_id=4019