Surface errors from external `importjs` call better
If the call to `importjs` fails, we would still try to parse the output
as json. We can detect failures using `v:shell_error` (which returns the
exit code from the last command) and show the output from the command as
an error string instead.

There's still an issue with "^@" showing up in the output, but I figured
this is at least better than what we had before.