Use importjsd background process
There is now an importjsd binary as an alternative to importjs. The new
tool will keep the process open and wait for input (using stdin). It
uses the jobs feature that was added in a (fairly) recent version of
vim:
https://groups.google.com/forum/#!msg/vim_dev/qtw5LdCrHJQ/uDhLZslbDQAJ

Consider this a POC. I'm still trying out a few things, expect
follow-ups.