Add check for empty resultString
I've noticed that if I run importjs very quickly after starting vim,
then the resultString is empty. This has been producing weird and
confusing error messages down the line, because the json_decode seems to
not error out, so result ends up not being a dictionary like we expect
it to.

To prevent this confusing scenario, I am just adding a check for this
case.