Add retry logic
If you try to run importjs very shortly after starting Vim, importjsd
hasn't had enough time to start up properly yet. This results in the
importjs command giving an empty response, which ends up causing errors.

Instead of erroring out, I decided that it would be nicer to simply
retry a couple of times. On my laptop, it seems to pick it up on the
second try, so I kept the number of tries pretty low.

This is a much better experience than erroring out confusingly. However,
I'm left wondering why we have an empty result instead of an error or
something.