Remove E715 error catching
In 5b0315cb, I scoped the try/catch to just the execution of the
importjs call. When reading this code, I originally thought that thi
E715 error was intended to apply to the json_decode, but I don't see how
that would possibly throw E715 since it expects a string, not a
dictionary, whereas ch_evalraw can accept a dictionary as its third
argument.

Since we are not using this third argument, it seems unlikely that this
error will actually happen here. I believe that it was likely misplaced
originally. In the meantime, I've improved edge case handling enough
where I don't think we actually need it anymore.