Adjust cursor based on file line count difference
When replacing the buffer via `importjs#ReplaceBuffer`, the original
cursor position is restored. However, the file has moved out from
underneath the cursor!

It is a better developer user experience to leave their cursor on the
line of code they started on, which can be calculated by diffing the
line counts and adding it to the original cursor position.