Use absolute path for pathToFile
I came across a bug when importing in a repo with multiple root
packages. The relative path used for the moduleName was way off. I
checked with other editor plugins for import-js and the bug wasn't
present. It turns out to be caused by using relative paths in the
payload sent to the import-js daemon. By making it absolute (e.g.
/Users/foo/project/file.js), we get the right relative paths back from
import-js.