Clear prompt in between unresolved imports
If you have multiple unresolved imports, you will be prompted to pick
the module you want for each of them. However, we weren't clearing out
the prompt in between, so it would end up just getting stacked on the
end, resulting in a garbled prompt.

To avoid this, I am making the plugin print an empty message immediately
before printing the inputlist.

Fixes #2