Highlight the unresolved import currently being prompted for
I want to have a little more context when resolving unresolved imports.
I think it will be helpful to highlight word that is currently being
prompted for. To ensure that this matches "Word" and not "WordFoo" I add
the special word-boundary characters to the matchadd call.

I'd like to also jump to the line that has the word to ensure that it
can be seen, but making sure the cursor ends up back where it belongs
may take a little work so I'm going to get this in as is for now.