Jump to word currently being prompted for
To help give more context when resolving unresolved imports, I recently
added highlighting of the current word that needs to be resolved. This
commit takes this a step farther by jumping to the line that has the
unresolved word.

It might be nice to center the line vertically in the buffer when we
jump, but for now I'm going to leave this as is and see how it feels.