Use default priority when highlighting unresolved word
I originally set this to -1, but after using it for a bit I noticed that
the syntax highlighting from Ale makes this setting entirely
ineffective. Since this is more useful information at the time of
resolving unresolved imports, I'm removing the -1 priority to let it use
the default priority, which is 10.

It's possible that we will want to increase this at some point, but for
now let's see how well the default works.