Lazy-load background importjs process
Instead of starting an importjs process on vim startup, we can wait
until someone issues the first importjs related command. This will
improve vim startup times and use less memory.

Thanks to @dylan-chong for pushing me in this direction. I had been
meaning to do it for a while but never got around to it.

Fixes #41